refactor/ line 관련 함수들 utils로 분리
